refactor scripts into modules and update tests
