+++ index.html
@@ -157,7 +157,11 @@
   </section>
 
   <!-- JavaScript -->
+  <script src="scripts/storage.js"></script>
+  <script src="scripts/notifications.js"></script>
+  <script src="scripts/admin.js"></script>
   <script src="scripts/csvParser.js"></script>
+  <script src="scripts/records.js"></script>
   <script src="scripts/app.js"></script>
 </body>
 </html>
