+++ scripts/app.js
@@ -1,17 +1,4 @@
 /* ---------- Initialization ---------- */
-function loadFromStorage(key, fallback) {
-  const item = localStorage.getItem(key);
-  if (!item) return fallback;
-  try {
-    return JSON.parse(item);
-  } catch (e) {
-    console.warn(`Failed to parse ${key} from storage, resetting to defaults`, e);
-    setTimeout(() => showError(`Stored data for ${key} was invalid and has been reset.`), 0);
-    saveToStorage(key, fallback);
-    return fallback;
-  }
-}
-
 let employees = loadFromStorage('employees', {});
 let equipmentItems = loadFromStorage('equipmentItems', {});
 let records = loadFromStorage('records', []);
@@ -23,104 +10,10 @@ let equipmentPage = 0;
 let employeeFilter = '';
 let equipmentFilter = '';
 
-function saveToStorage(key, data) {
-  localStorage.setItem(key, JSON.stringify(data));
-}
-
-function csvEscape(value) {
-  return String(value).replace(/"/g, '""');
-}
-
-/* ---------- Notification System ---------- */
-let notificationTimer;
-let tempNotificationActive = false;
-
-function clearNotification() {
-  const notificationDiv = document.getElementById('notifications');
-  notificationDiv.classList.remove('visible');
-  notificationDiv.className = '';
-  notificationDiv.textContent = '';
-  tempNotificationActive = false;
-  updateNotifications();
-}
-
-function showNotification(message, type, delay = 3000) {
-  const notificationDiv = document.getElementById('notifications');
-  if (notificationTimer) {
-    clearTimeout(notificationTimer);
-  }
-
-  tempNotificationActive = true;
-  notificationDiv.className = type;
-  notificationDiv.textContent = message;
-  notificationDiv.classList.add('visible');
-
-  if (delay > 0) {
-    notificationTimer = setTimeout(clearNotification, delay);
-  } else {
-    clearNotification();
-  }
-}
-
-function showSuccess(message, delay) {
-  showNotification(message, 'success', delay);
-}
-
-function showError(message, delay) {
-  showNotification(message, 'error', delay);
-}
-
-function setFieldError(input, message) {
-  const errorSpan = input.parentElement.querySelector('.error-message');
-  if (errorSpan) {
-    errorSpan.textContent = message;
-  }
-  input.classList.add('error');
-  input.setAttribute('aria-invalid', 'true');
-}
-
-function clearFieldError(input) {
-  const errorSpan = input.parentElement.querySelector('.error-message');
-  if (errorSpan) {
-    errorSpan.textContent = '';
-  }
-  input.classList.remove('error');
-  input.removeAttribute('aria-invalid');
-}
-
-function updateNotifications() {
-  if (tempNotificationActive) return;
-
-  const notificationDiv = document.getElementById('notifications');
-  notificationDiv.className = '';
-  notificationDiv.classList.remove('visible');
-  const status = {};
-  records.forEach(rec => {
-    (rec.equipmentBarcodes || []).forEach(code => {
-      if (!status[code]) status[code] = 0;
-      if (rec.action === "Check-Out") {
-        status[code]++;
-      } else if (rec.action === "Check-In") {
-        status[code]--;
-      }
-    });
-  });
-  const overdue = [];
-  for (let code in status) {
-    if (status[code] > 0) {
-      const name = equipmentItems[code] || "Unknown Equipment";
-      overdue.push(`${code} (${name})`);
-    }
-  }
-  if (overdue.length > 0) {
-    notificationDiv.textContent = "Overdue Equipment: " + overdue.join(", ");
-    notificationDiv.classList.add('visible');
-  } else {
-    notificationDiv.textContent = "";
-    notificationDiv.classList.remove('visible');
-  }
-}
 updateNotifications();
+
+const actionBtn = document.getElementById('actionBtn');
+const actionMenu = document.getElementById('actionMenu');
 
 /* ---------- Navigation ---------- */
 function showSection(section) {
@@ -204,139 +97,107 @@ function removeEquipmentField(button) {
 }
 
 function updateRemoveButtons() {
-  const items = document.querySelectorAll('#equipmentList .equipment-item');
-  items.forEach(item => {
-    const btn = item.querySelector('.removeEquipment');
-    if (items.length > 1) {
-      btn.classList.remove('hidden');
+  const equipmentItems = document.querySelectorAll('#equipmentList .equipment-item');
+  equipmentItems.forEach(item => {
+    const input = item.querySelector('input[name="equipment"]');
+    const removeBtn = item.querySelector('.removeEquipment');
+    if (input.disabled) {
+      removeBtn.classList.add('hidden');
     } else {
-      btn.classList.add('hidden');
+      removeBtn.classList.remove('hidden');
     }
   });
 }
 
 function lookupEmployee() {
-  const input = document.getElementById('badge');
-  clearFieldError(input);
-  const badge = input.value.trim();
-  const employeeNameSpan = document.getElementById('employeeName');
+  const badgeInput = document.getElementById('badge');
+  const badge = badgeInput.value.trim();
+  const nameDisplay = document.getElementById('employeeName');
+  clearFieldError(badgeInput);
+  if (badge === '') {
+    nameDisplay.textContent = '';
+    return;
+  }
   if (employees[badge]) {
-    employeeNameSpan.textContent = employees[badge];
-    const equipmentInput = document.querySelector('input[name="equipment"]:not([disabled])');
-    if (equipmentInput) {
-      equipmentInput.focus();
-    }
+    nameDisplay.textContent = employees[badge];
   } else {
-    employeeNameSpan.textContent = "";
+    nameDisplay.textContent = 'Unknown employee';
   }
 }
 
 function lookupEquipment(event) {
   const input = event.target;
+  const code = input.value.trim();
+  const nameDisplay = input.parentElement.querySelector('.equipmentNameDisplay');
   clearFieldError(input);
-  const value = input.value.trim();
-  const display = input.parentElement.querySelector('.equipmentNameDisplay');
-  const equipmentName = equipmentItems[value];
-  display.textContent = equipmentName || "";
-
-  if (equipmentName) {
-    const equipmentList = document.getElementById('equipmentList');
-    const inputs = equipmentList.querySelectorAll('input[name="equipment"]');
-
-    for (const other of inputs) {
-      if (other !== input && other.value.trim() === value) {
-        setFieldError(input, 'Equipment barcode already entered.');
-        showError('Error: Equipment barcode already entered.');
-        input.value = '';
-        display.textContent = '';
-        return;
-      }
-    }
-
-    input.disabled = true;
-    if (inputs[inputs.length - 1] === input) {
-      addEquipmentField();
-      equipmentList.lastElementChild.querySelector('input[name="equipment"]').focus();
-    }
+  if (code === '') {
+    nameDisplay.textContent = '';
+    input.disabled = false;
+    updateRemoveButtons();
+    return;
   }
+  const existing = Array.from(document.querySelectorAll('#equipmentList input[name="equipment"]'))
+    .filter(el => el !== input && el.value.trim() === code);
+  if (existing.length > 0) {
+    showError('Duplicate equipment barcode!');
+    input.value = '';
+    input.disabled = false;
+    nameDisplay.textContent = '';
+    updateRemoveButtons();
+    return;
+  }
+  if (equipmentItems[code]) {
+    nameDisplay.textContent = equipmentItems[code];
+    input.disabled = true;
+  } else {
+    nameDisplay.textContent = 'Unknown equipment';
+    input.disabled = true;
+  }
+  if (document.querySelectorAll('#equipmentList input[name="equipment"]:not([disabled])').length === 0) {
+    addEquipmentField();
+  }
+  updateRemoveButtons();
 }
 
 document.getElementById('checkoutForm').addEventListener('submit', function(e) {
   e.preventDefault();
-
   const badgeInput = document.getElementById('badge');
+  const actionInput = document.getElementById('action');
   clearFieldError(badgeInput);
+  clearFieldError(actionInput);
   const badge = badgeInput.value.trim();
+  const action = actionInput.value.trim();
+  let hasError = false;
   if (!badge) {
     setFieldError(badgeInput, 'Badge ID is required.');
-    return;
+    hasError = true;
   }
-  if (!employees[badge]) {
-    setFieldError(badgeInput, 'Employee badge not recognized.');
-    return;
-  }
-
-  const equipmentInputs = Array.from(document.querySelectorAll('#equipmentList input[name="equipment"]'));
-  equipmentInputs.forEach(clearFieldError);
-  const equipmentCodes = equipmentInputs
-    .map(input => input.value.trim())
-    .filter(code => code !== "");
-
-  if (equipmentCodes.length === 0) {
-    setFieldError(equipmentInputs[0], 'Please scan at least one equipment barcode.');
-    return;
-  }
-
-  for (const input of equipmentInputs) {
-    const code = input.value.trim();
-    if (code && !equipmentItems[code]) {
-      setFieldError(input, "Equipment barcode '" + code + "' not recognized.");
-      return;
-    }
-  }
-
-  const employeeName = employees[badge] || "Unknown";
-  const actionInput = document.getElementById('action');
-  const action = actionInput.value;
   if (!action) {
-    showError('Please select an action.');
-    return;
+    setFieldError(actionInput, 'Please select an action.');
+    hasError = true;
   }
-  const status = {};
-  records.forEach(rec => {
-    (rec.equipmentBarcodes || []).forEach(code => {
-      if (!status[code]) status[code] = 0;
-      if (rec.action === "Check-Out") {
-        status[code]++;
-      } else if (rec.action === "Check-In") {
-        status[code]--;
-      }
-    });
-  });
-  if (action === "Check-Out") {
-    for (const code of equipmentCodes) {
-      if (status[code] > 0) {
-        const offendingInput = equipmentInputs.find(input => input.value.trim() === code);
-        if (offendingInput) {
-          setFieldError(offendingInput, "Equipment barcode '" + code + "' is already checked out.");
-        }
-        showError("Equipment barcode '" + code + "' is already checked out.");
-        return;
-      }
+  if (hasError) return;
+  const equipmentInputs = document.querySelectorAll('#equipmentList input[name="equipment"]');
+  const equipmentBarcodes = [];
+  const equipmentNamesList = [];
+  equipmentInputs.forEach(input => {
+    const code = input.value.trim();
+    if (code) {
+      equipmentBarcodes.push(code);
+      equipmentNamesList.push(equipmentItems[code] || 'Unknown equipment');
     }
-  }
-  const now = new Date();
-  const timeString = now.toLocaleTimeString('en-US', {hour: '2-digit', minute: '2-digit', second: '2-digit', hour12: true});
-  const recordDate = now.toISOString().substring(0,10);
-  const timestamp = `${(now.getMonth()+1).toString().padStart(2, '0')}/${now.getDate().toString().padStart(2, '0')}/${now.getFullYear()} ${timeString}`;
-
-  const equipmentBarcodes = equipmentCodes;
-  const equipmentNames = equipmentCodes.map(code => equipmentItems[code] || "");
-
-  const record = { timestamp, recordDate, badge, employeeName, equipmentBarcodes, equipmentNames, action };
+  });
+  const record = {
+    timestamp: new Date().toISOString(),
+    recordDate: new Date().toISOString().substring(0,10),
+    badge: badge,
+    employeeName: employees[badge] || 'Unknown employee',
+    equipmentBarcodes: equipmentBarcodes,
+    equipmentNames: equipmentNamesList,
+    action: action
+  };
   records.push(record);
   saveToStorage('records', records);
-  updateNotifications();
   showSuccess('Record saved locally!');
   this.reset();
   document.getElementById('employeeName').textContent = "";
@@ -347,415 +208,6 @@ document.getElementById('checkoutForm').addEventListener('submit', function(e) {
   addEquipmentField();
   document.getElementById('badge').focus();
 });
-
-/* ---------- Admin Panel Functions ---------- */
-function displayEmployeeList(page = employeePage, filter = employeeFilter) {
-  filter = (filter || '').toLowerCase();
-  const list = document.getElementById('employeeList');
-  list.innerHTML = "";
-  const entries = Object.entries(employees).sort((a, b) => a[0].localeCompare(b[0]));
-  const filtered = entries.filter(([badge, name]) =>
-    badge.toLowerCase().includes(filter) || name.toLowerCase().includes(filter)
-  );
-  const totalPages = Math.max(1, Math.ceil(filtered.length / pageSize));
-  page = Math.min(Math.max(page, 0), totalPages - 1);
-  const start = page * pageSize;
-  const pageItems = filtered.slice(start, start + pageSize);
-  if (pageItems.length === 0) {
-    const placeholder = document.createElement('li');
-    placeholder.className = 'placeholder';
-    placeholder.textContent = 'No employees added yet.';
-    list.appendChild(placeholder);
-  } else {
-    pageItems.forEach(([badge, name]) => {
-      const li = document.createElement('li');
-      const textSpan = document.createElement('span');
-      textSpan.textContent = `${badge}: ${name}`;
-      const del = document.createElement('button');
-      del.type = 'button';
-      del.className = 'deleteEmployee';
-      del.textContent = '❌';
-      del.title = 'Remove Employee';
-      del.setAttribute('aria-label', 'Remove Employee');
-      del.addEventListener('click', () => removeEmployee(badge));
-      li.appendChild(textSpan);
-      li.appendChild(del);
-      list.appendChild(li);
-    });
-  }
-  employeePage = page;
-  employeeFilter = filter;
-  const prevBtn = document.getElementById('employeePrev');
-  const nextBtn = document.getElementById('employeeNext');
-  const pageIndicator = document.getElementById('employeePageIndicator');
-  if (prevBtn) {
-    const hidePrev = page <= 0;
-    prevBtn.classList.toggle('hidden', hidePrev);
-    prevBtn.disabled = hidePrev;
-  }
-  if (nextBtn) {
-    const hideNext = page >= totalPages - 1;
-    nextBtn.classList.toggle('hidden', hideNext);
-    nextBtn.disabled = hideNext;
-  }
-  if (pageIndicator) pageIndicator.textContent = `Page ${page + 1} of ${totalPages}`;
-}
-function addEmployee() {
-  const nameInput = document.getElementById('empName');
-  const badgeInput = document.getElementById('empBadge');
-  clearFieldError(nameInput);
-  clearFieldError(badgeInput);
-  const badge = badgeInput.value.trim();
-  const name = nameInput.value.trim();
-  let hasError = false;
-  if (!name) {
-    setFieldError(nameInput, 'Employee name is required.');
-    hasError = true;
-  }
-  if (!badge) {
-    setFieldError(badgeInput, 'Badge ID is required.');
-    hasError = true;
-  }
-  if (hasError) return;
-  if (employees[badge]) {
-    setFieldError(badgeInput, 'Badge ID already exists.');
-    showError('Employee with this badge ID already exists!');
-    return;
-  }
-  employees[badge] = name;
-  saveToStorage('employees', employees);
-  showSuccess('Employee added successfully!');
-  displayEmployeeList();
-  document.getElementById('adminForm').reset();
-  clearFieldError(nameInput);
-  clearFieldError(badgeInput);
-}
-function removeEmployee(badge) {
-  if (!badge || !employees[badge]) {
-    showError('Invalid badge ID or employee not found!');
-    return;
-  }
-  const confirmed = typeof confirm === 'function'
-    ? confirm('Are you sure you want to remove this employee?')
-    : true;
-  if (!confirmed) return;
-  delete employees[badge];
-  saveToStorage('employees', employees);
-  showSuccess('Employee removed successfully!');
-  displayEmployeeList();
-}
-function displayEquipmentListAdmin(page = equipmentPage, filter = equipmentFilter) {
-  filter = (filter || '').toLowerCase();
-  const list = document.getElementById('equipmentListAdmin');
-  list.innerHTML = "";
-  const entries = Object.entries(equipmentItems).sort((a, b) => a[0].localeCompare(b[0]));
-  const filtered = entries.filter(([serial, name]) =>
-    serial.toLowerCase().includes(filter) || name.toLowerCase().includes(filter)
-  );
-  const totalPages = Math.max(1, Math.ceil(filtered.length / pageSize));
-  page = Math.min(Math.max(page, 0), totalPages - 1);
-  const start = page * pageSize;
-  const pageItems = filtered.slice(start, start + pageSize);
-  if (pageItems.length === 0) {
-    const placeholder = document.createElement('li');
-    placeholder.className = 'placeholder';
-    placeholder.textContent = 'No equipment added yet.';
-    list.appendChild(placeholder);
-  } else {
-    pageItems.forEach(([serial, name]) => {
-      const li = document.createElement('li');
-      const textSpan = document.createElement('span');
-      textSpan.textContent = `${serial}: ${name}`;
-      const del = document.createElement('button');
-      del.type = 'button';
-      del.className = 'deleteEquipment';
-      del.textContent = '❌';
-      del.title = 'Remove Equipment';
-      del.setAttribute('aria-label', 'Remove Equipment');
-      del.addEventListener('click', () => removeEquipmentAdmin(serial));
-      li.appendChild(textSpan);
-      li.appendChild(del);
-      list.appendChild(li);
-    });
-  }
-  equipmentPage = page;
-  equipmentFilter = filter;
-  const prevBtn = document.getElementById('equipmentPrev');
-  const nextBtn = document.getElementById('equipmentNext');
-  const pageIndicator = document.getElementById('equipmentPageIndicator');
-  if (prevBtn) {
-    const hidePrev = page <= 0;
-    prevBtn.classList.toggle('hidden', hidePrev);
-    prevBtn.disabled = hidePrev;
-  }
-  if (nextBtn) {
-    const hideNext = page >= totalPages - 1;
-    nextBtn.classList.toggle('hidden', hideNext);
-    nextBtn.disabled = hideNext;
-  }
-  if (pageIndicator) pageIndicator.textContent = `Page ${page + 1} of ${totalPages}`;
-}
-function addEquipmentAdmin() {
-  const nameInput = document.getElementById('equipName');
-  const serialInput = document.getElementById('equipSerial');
-  clearFieldError(nameInput);
-  clearFieldError(serialInput);
-  const serial = serialInput.value.trim();
-  const name = nameInput.value.trim();
-  let hasError = false;
-  if (!name) {
-    setFieldError(nameInput, 'Equipment name is required.');
-    hasError = true;
-  }
-  if (!serial) {
-    setFieldError(serialInput, 'Equipment serial is required.');
-    hasError = true;
-  }
-  if (hasError) return;
-  if (equipmentItems[serial]) {
-    setFieldError(serialInput, 'Equipment serial already exists.');
-    showError('Equipment with this serial already exists!');
-    return;
-  }
-  equipmentItems[serial] = name;
-  saveToStorage('equipmentItems', equipmentItems);
-  showSuccess('Equipment added successfully!');
-  displayEquipmentListAdmin();
-  document.getElementById('equipmentAdminForm').reset();
-  clearFieldError(nameInput);
-  clearFieldError(serialInput);
-}
-function removeEquipmentAdmin(serial) {
-  if (!serial || !equipmentItems[serial]) {
-    showError('Invalid equipment serial or equipment not found!');
-    return;
-  }
-  const confirmed = typeof confirm === 'function'
-    ? confirm('Are you sure you want to remove this equipment?')
-    : true;
-  if (!confirmed) return;
-  delete equipmentItems[serial];
-  saveToStorage('equipmentItems', equipmentItems);
-  showSuccess('Equipment removed successfully!');
-  displayEquipmentListAdmin();
-}
-
-/* ---------- Record Filtering & Export ---------- */
-// Escape HTML entities to prevent script injection in dynamic table
-function escapeHtml(str) {
-  return String(str)
-    .replace(/&/g, '&amp;')
-    .replace(/</g, '&lt;')
-    .replace(/>/g, '&gt;')
-    .replace(/'/g, '&#39;')
-    .replace(/"/g, '&quot;');
-}
-
-function displayRecords(recArray) {
-  const container = document.getElementById('recordsTable');
-  if (!recArray.length) {
-    container.innerHTML = "<p>No records found.</p>";
-    return;
-  }
-  let html = "<table><caption>Equipment check-in/out records</caption><tr><th>Timestamp</th><th>Badge</th><th>Name</th><th>Equipment Barcodes</th><th>Equipment Names</th><th>Action</th></tr>";
-  recArray.forEach(rec => {
-    // Escape record fields before injecting into the table
-    html += `<tr>
-          <td>${escapeHtml(rec.timestamp)}</td>
-          <td>${escapeHtml(rec.badge)}</td>
-          <td>${escapeHtml(rec.employeeName)}</td>
-          <td>${escapeHtml((rec.equipmentBarcodes ?? []).join('; '))}</td>
-          <td>${escapeHtml((rec.equipmentNames ?? []).join('; '))}</td>
-          <td>${escapeHtml(rec.action)}</td>
-        </tr>`;
-  });
-  html += "</table>";
-  container.innerHTML = html;
-}
-function filterRecords() {
-  const search = document.getElementById('recordSearch').value.trim().toLowerCase();
-  const equipSearch = document.getElementById('recordEquipment').value.trim().toLowerCase();
-  const date = document.getElementById('recordDate').value;
-  let filtered = records;
-  if (search) {
-    filtered = filtered.filter(rec => {
-      const badge = String(rec?.badge ?? "");
-      const name = String(rec?.employeeName ?? "");
-      return badge.toLowerCase().includes(search) ||
-             name.toLowerCase().includes(search);
-    });
-  }
-  if (equipSearch) {
-    filtered = filtered.filter(rec => {
-      const barcodes = (rec?.equipmentBarcodes ?? []).join(" ").toLowerCase();
-      const names = (rec?.equipmentNames ?? []).join(" ").toLowerCase();
-      const combinedEquip = `${barcodes} ${names}`.trim();
-      return combinedEquip.includes(equipSearch);
-    });
-  }
-  if (date) {
-    filtered = filtered.filter(rec => rec.recordDate === date);
-  }
-  displayRecords(filtered);
-}
-function clearFilters() {
-  document.getElementById('recordSearch').value = "";
-  document.getElementById('recordEquipment').value = "";
-  document.getElementById('recordDate').value = "";
-  displayRecords(records);
-}
-function exportRecordsCSV() {
-  if (!records.length) {
-    showError("No records to export.");
-    return;
-  }
-  const header = "Timestamp,Employee Badge ID,Employee Name,Equipment Barcodes,Equipment Names,Action\n";
-  const rows = records.map(rec =>
-    `"${csvEscape(rec.timestamp ?? '')}",` +
-    `"${csvEscape(rec.badge ?? '')}",` +
-    `"${csvEscape(rec.employeeName ?? '')}",` +
-    `"${csvEscape((rec.equipmentBarcodes ?? []).join('; ') ?? '')}",` +
-    `"${csvEscape((rec.equipmentNames ?? []).join('; ') ?? '')}",` +
-    `"${csvEscape(rec.action ?? '')}"`
-  );
-  const csvContent = 'data:text/csv;charset=utf-8,' + header + rows.join("\n");
-  const link = document.createElement('a');
-  link.href = encodeURI(csvContent);
-  link.download = `Records_${new Date().toISOString().substring(0,10)}.csv`;
-  document.body.appendChild(link);
-  link.click();
-  document.body.removeChild(link);
-}
-
-/* ---------- CSV Import/Export Functions ---------- */
-function exportEmployeesCSV() {
-  let csvContent = "data:text/csv;charset=utf-8,";
-  csvContent += "Badge ID,Employee Name\n";
-  Object.entries(employees).forEach(([badge, name]) => {
-    csvContent += `"${csvEscape(badge)}","${csvEscape(name)}"\n`;
-  });
-  const link = document.createElement("a");
-  link.href = encodeURI(csvContent);
-  link.download = "employees.csv";
-  document.body.appendChild(link);
-  link.click();
-  document.body.removeChild(link);
-}
-
-function exportEquipmentCSV() {
-  let csvContent = "data:text/csv;charset=utf-8,";
-  csvContent += "Equipment Serial,Equipment Name\n";
-  Object.entries(equipmentItems).forEach(([serial, name]) => {
-    csvContent += `"${csvEscape(serial)}","${csvEscape(name)}"\n`;
-  });
-  const link = document.createElement("a");
-  link.href = encodeURI(csvContent);
-  link.download = "equipment.csv";
-  document.body.appendChild(link);
-  link.click();
-  document.body.removeChild(link);
-}
-
-function triggerImportEmployees() {
-  document.getElementById("importEmployeesFile").click();
-}
-
-function setLoading(button, loading, text = 'Importing...') {
-  if (!button) return;
-  if (loading) {
-    button.disabled = true;
-    button.dataset.originalHtml = button.innerHTML;
-    button.textContent = text;
-  } else {
-    button.disabled = false;
-    if (button.dataset.originalHtml) {
-      button.innerHTML = button.dataset.originalHtml;
-      delete button.dataset.originalHtml;
-    }
-  }
-}
-
-function handleImportEmployees(event) {
-  const input = event.target;
-  const button = document.getElementById('importEmployeesAction');
-  setLoading(button, true);
-  if (input.files.length === 0) {
-    setLoading(button, false);
-    return;
-  }
-  const file = input.files[0];
-  const reader = new FileReader();
-  reader.onerror = function() {
-    showError("Unable to read employees CSV file.");
-    setLoading(button, false);
-    input.value = "";
-  };
-  reader.onload = function(e) {
-    const text = e.target.result;
-    const rows = parseCSV(text);
-    for (let i = 1; i < rows.length; i++) {
-      const parts = rows[i];
-      if (parts.length < 2) {
-        showError(`Skipping malformed line ${i + 1}: ${parts.join(',')}`);
-        continue;
-      }
-      let badge = parts[0].replace(/^"|"$/g, '').trim();
-      let name = parts[1].replace(/^"|"$/g, '').trim();
-      if (badge && name) {
-        employees[badge] = name;
-      }
-    }
-    saveToStorage("employees", employees);
-    showSuccess("Employee CSV import completed successfully.");
-    displayEmployeeList();
-    setLoading(button, false);
-    input.value = "";
-  };
-  reader.readAsText(file);
-}
-
-function triggerImportEquipment() {
-  document.getElementById("importEquipmentFile").click();
-}
-
-function handleImportEquipment(event) {
-  const input = event.target;
-  const button = document.getElementById('importEquipmentAction');
-  setLoading(button, true);
-  if (input.files.length === 0) {
-    setLoading(button, false);
-    return;
-  }
-  const file = input.files[0];
-  const reader = new FileReader();
-  reader.onerror = function() {
-    showError("Unable to read equipment CSV file.");
-    setLoading(button, false);
-    input.value = "";
-  };
-  reader.onload = function(e) {
-    const text = e.target.result;
-    const rows = parseCSV(text);
-    for (let i = 1; i < rows.length; i++) {
-      const parts = rows[i];
-      if (parts.length < 2) {
-        showError(`Skipping malformed line ${i + 1}: ${parts.join(',')}`);
-        continue;
-      }
-      let serial = parts[0].replace(/^"|"$/g, '').trim();
-      let name = parts[1].replace(/^"|"$/g, '').trim();
-      if (serial && name) {
-        equipmentItems[serial] = name;
-      }
-    }
-    saveToStorage("equipmentItems", equipmentItems);
-    showSuccess("Equipment CSV import completed successfully.");
-    displayEquipmentListAdmin();
-    setLoading(button, false);
-    input.value = "";
-  };
-  reader.readAsText(file);
-}
 
 /* ---------- Event Listeners ---------- */
 
@@ -800,6 +252,14 @@ document.getElementById('equipmentAdminForm').addEventListener('submit', (e) => 
 
 function setupDropdown(button, menu, onSelect) {
   const items = Array.from(menu.querySelectorAll('button'));
+  items.forEach((item, idx) => {
+    item.addEventListener('keydown', (e) => {
+      if (e.key === 'Tab' && !e.shiftKey && idx === items.length - 1) {
+        e.preventDefault();
+        closeMenu();
+      }
+    });
+  });
 
   const openMenu = () => {
     menu.classList.remove('hidden');
@@ -841,110 +301,32 @@ function setupDropdown(button, menu, onSelect) {
     }
   });
 
-  items.forEach((item, index) => {
-    item.setAttribute('tabindex', '-1');
-    item.addEventListener('click', () => {
-      onSelect?.(item);
-      closeMenu();
-    });
-    item.addEventListener('keydown', (e) => {
-      if (e.key === 'ArrowDown') {
-        e.preventDefault();
-        items[(index + 1) % items.length].focus();
-      } else if (e.key === 'ArrowUp') {
-        e.preventDefault();
-        items[(index - 1 + items.length) % items.length].focus();
-      } else if (e.key === 'Escape') {
-        e.preventDefault();
-        closeMenu();
-      } else if (e.key === 'Tab') {
-        e.preventDefault();
-        closeMenu();
-      }
-    });
+  menu.addEventListener('click', (e) => {
+    const item = e.target.closest('button');
+    if (!item) return;
+    onSelect(item.dataset.value, item.textContent);
+    closeMenu();
   });
 
-  document.addEventListener('click', (e) => {
-    if (!menu.contains(e.target) && e.target !== button) {
-      if (button.getAttribute('aria-expanded') === 'true') {
-        closeMenu();
-      }
+  menu.addEventListener('keydown', (e) => {
+    const currentIndex = items.indexOf(document.activeElement);
+    if (e.key === 'Tab' && !e.shiftKey && currentIndex === items.length - 1) {
+      e.preventDefault();
+      closeMenu();
+    } else if (e.key === 'ArrowDown') {
+      e.preventDefault();
+      items[(currentIndex + 1) % items.length]?.focus();
+    } else if (e.key === 'ArrowUp') {
+      e.preventDefault();
+      items[(currentIndex - 1 + items.length) % items.length]?.focus();
+    } else if (e.key === 'Escape') {
+      e.preventDefault();
+      closeMenu();
     }
   });
 }
 
-const actionBtn = document.getElementById('actionBtn');
-const actionMenu = document.getElementById('actionMenu');
-if (actionBtn && actionMenu) {
-  setupDropdown(actionBtn, actionMenu, (item) => {
-    document.getElementById('action').value = item.dataset.value || item.textContent;
-    actionBtn.textContent = item.textContent;
-  });
-}
-
-const importExportBtn = document.getElementById('importExportBtn');
-const importExportMenu = document.getElementById('importExportMenu');
-if (importExportBtn && importExportMenu) {
-  setupDropdown(importExportBtn, importExportMenu);
-}
-
-document.addEventListener('click', (e) => {
-  const clickedInsideDropdown =
-    actionMenu?.contains(e.target) ||
-    actionBtn?.contains(e.target) ||
-    importExportMenu?.contains(e.target) ||
-    importExportBtn?.contains(e.target);
-
-  if (!clickedInsideDropdown) {
-    actionMenu?.classList.add('hidden');
-    actionBtn?.setAttribute('aria-expanded', 'false');
-    importExportMenu?.classList.add('hidden');
-    importExportBtn?.setAttribute('aria-expanded', 'false');
-  }
+setupDropdown(actionBtn, actionMenu, (value, text) => {
+  document.getElementById('action').value = value;
+  actionBtn.textContent = text;
 });
-
-document.getElementById('exportEmployeesAction').addEventListener('click', () => {
-  exportEmployeesCSV();
-});
-document.getElementById('importEmployeesAction').addEventListener('click', () => {
-  triggerImportEmployees();
-});
-document.getElementById('exportEquipmentAction').addEventListener('click', () => {
-  exportEquipmentCSV();
-});
-document.getElementById('importEquipmentAction').addEventListener('click', () => {
-  triggerImportEquipment();
-});
-document.getElementById('importEmployeesFile').addEventListener('change', handleImportEmployees);
-document.getElementById('importEquipmentFile').addEventListener('change', handleImportEquipment);
-
-document.getElementById('employeeSearch').addEventListener('input', (e) => {
-  displayEmployeeList(0, e.target.value.trim().toLowerCase());
-});
-document.getElementById('employeePrev').addEventListener('click', () => {
-  displayEmployeeList(employeePage - 1);
-});
-document.getElementById('employeeNext').addEventListener('click', () => {
-  displayEmployeeList(employeePage + 1);
-});
-
-document.getElementById('equipmentSearch').addEventListener('input', (e) => {
-  displayEquipmentListAdmin(0, e.target.value.trim().toLowerCase());
-});
-document.getElementById('equipmentPrev').addEventListener('click', () => {
-  displayEquipmentListAdmin(equipmentPage - 1);
-});
-document.getElementById('equipmentNext').addEventListener('click', () => {
-  displayEquipmentListAdmin(equipmentPage + 1);
-});
-
-document.getElementById('recordFilterForm').addEventListener('submit', (e) => {
-  e.preventDefault();
-  filterRecords();
-});
-document.getElementById('clearFiltersBtn').addEventListener('click', clearFilters);
-document.getElementById('exportRecordsBtn').addEventListener('click', exportRecordsCSV);
-
-// Set initial active section
-showSection('checkout');
-
