+++ tests/importHandlers.test.js
@@ -3,8 +3,14 @@ const path = require('path');
 const { JSDOM } = require('jsdom');
 
 const html = fs.readFileSync(path.resolve(__dirname, '../index.html'), 'utf8');
-const parserScript = fs.readFileSync(path.resolve(__dirname, '../scripts/csvParser.js'), 'utf8');
-const appScript = fs.readFileSync(path.resolve(__dirname, '../scripts/app.js'), 'utf8');
+const scripts = [
+  '../scripts/storage.js',
+  '../scripts/notifications.js',
+  '../scripts/admin.js',
+  '../scripts/csvParser.js',
+  '../scripts/records.js',
+  '../scripts/app.js'
+].map(p => fs.readFileSync(path.resolve(__dirname, p), 'utf8')).join('\n');
 
 function setupDom() {
   const dom = new JSDOM(html, { url: 'http://localhost', runScripts: 'dangerously' });
@@ -12,8 +18,7 @@ function setupDom() {
   global.window = window;
   global.document = window.document;
   global.localStorage = window.localStorage;
-  window.eval(parserScript);
-  window.eval(appScript);
+  window.eval(scripts);
   return window;
 }
